remove welcomeMessage component; move and re-factor PromptList component
